+++ veilid-core/src/routing_table/tasks/bootstrap.rs
@@ -160,6 +160,8 @@ impl RoutingTable {
             }
         }
 
+        self.flush().await;
+
         Ok(())
     }
 }